+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `engine='mesh'` for `AdjacencyGraph.save_surface_obj` for mesh output (https://github.com/chenzhaiyu/abspy/issues/18).
+
+### Removed
+- `engine='projection'` and `engine='sorted'` for `AdjacencyGraph.save_surface_obj`.
+
+### Changed
+- Rename `engine='rendering'` &rarr; `engine='soup'` for `AdjacencyGraph.save_surface_obj`.
 
 ## [0.2.5] - 2024-12-01
 ### Added